requirements.txt
################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

boto3==1.10.26
botocore==1.13.26
certifi==2019.9.11
chardet==3.0.4
django-countries==5.5
django-dotenv==1.4.2
django-seed==0.1.9
django-storages==1.8
django==2.2.5
docutils==0.15.2
faker==2.0.4
idna==2.8
jmespath==0.9.4
pillow==6.2.1
python-dateutil==2.8.0 ; python_version >= '2.7'
pytz==2019.3
requests==2.22.0
s3transfer==0.2.1
sentry-sdk==0.13.2
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
psycopg2==2.8
urllib3==1.25.7 ; python_version >= '3.4'
